+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonManager.cs
@@ -1957,6 +1957,114 @@ public sealed class KryptonManager : Component
 
 
 
+    public static PaletteMicrosoft365LimeGreenDark PaletteMicrosoft365LimeGreenDark => _paletteMicrosoft365LimeGreenDark ??= new PaletteMicrosoft365LimeGreenDark();
+    public static PaletteOffice2007MaterializeBlue PaletteOffice2007MaterializeBlue => _paletteOffice2007MaterializeBlue ??= new PaletteOffice2007MaterializeBlue();
+    public static PaletteOffice2007MaterializeBlueDark PaletteOffice2007MaterializeBlueDark => _paletteOffice2007MaterializeBlueDark ??= new PaletteOffice2007MaterializeBlueDark();
+    public static PaletteOffice2007MaterializeLightBlue PaletteOffice2007MaterializeLightBlue => _paletteOffice2007MaterializeLightBlue ??= new PaletteOffice2007MaterializeLightBlue();
+    public static PaletteOffice2007MaterializeLightBlueDark PaletteOffice2007MaterializeLightBlueDark => _paletteOffice2007MaterializeLightBlueDark ??= new PaletteOffice2007MaterializeLightBlueDark();
+    public static PaletteOffice2007SilverDarkModeAlternate PaletteOffice2007SilverDarkModeAlternate => _paletteOffice2007SilverDarkModeAlternate ??= new PaletteOffice2007SilverDarkModeAlternate();
+    public static PaletteOffice2010MaterializeBlue PaletteOffice2010MaterializeBlue => _paletteOffice2010MaterializeBlue ??= new PaletteOffice2010MaterializeBlue();
+    public static PaletteOffice2010MaterializeBlueDark PaletteOffice2010MaterializeBlueDark => _paletteOffice2010MaterializeBlueDark ??= new PaletteOffice2010MaterializeBlueDark();
+    public static PaletteOffice2010MaterializeLightBlue PaletteOffice2010MaterializeLightBlue => _paletteOffice2010MaterializeLightBlue ??= new PaletteOffice2010MaterializeLightBlue();
+    public static PaletteOffice2010MaterializeLightBlueDark PaletteOffice2010MaterializeLightBlueDark => _paletteOffice2010MaterializeLightBlueDark ??= new PaletteOffice2010MaterializeLightBlueDark();
+    public static PaletteOffice2010SilverDarkModeAlternate PaletteOffice2010SilverDarkModeAlternate => _paletteOffice2010SilverDarkModeAlternate ??= new PaletteOffice2010SilverDarkModeAlternate();
+    public static PaletteOffice2013MaterializeBlue PaletteOffice2013MaterializeBlue => _paletteOffice2013MaterializeBlue ??= new PaletteOffice2013MaterializeBlue();
+    public static PaletteOffice2013MaterializeBlueDark PaletteOffice2013MaterializeBlueDark => _paletteOffice2013MaterializeBlueDark ??= new PaletteOffice2013MaterializeBlueDark();
+    public static PaletteOffice2013MaterializeLightBlue PaletteOffice2013MaterializeLightBlue => _paletteOffice2013MaterializeLightBlue ??= new PaletteOffice2013MaterializeLightBlue();
+    public static PaletteOffice2013MaterializeLightBlueDark PaletteOffice2013MaterializeLightBlueDark => _paletteOffice2013MaterializeLightBlueDark ??= new PaletteOffice2013MaterializeLightBlueDark();
+    public static PaletteOffice2013SilverDarkModeAlternate PaletteOffice2013SilverDarkModeAlternate => _paletteOffice2013SilverDarkModeAlternate ??= new PaletteOffice2013SilverDarkModeAlternate();
+    public static PaletteMicrosoft365MaterializeBlue PaletteMicrosoft365MaterializeBlue => _paletteMicrosoft365MaterializeBlue ??= new PaletteMicrosoft365MaterializeBlue();
+    public static PaletteMicrosoft365MaterializeBlueDark PaletteMicrosoft365MaterializeBlueDark => _paletteMicrosoft365MaterializeBlueDark ??= new PaletteMicrosoft365MaterializeBlueDark();
+    public static PaletteMicrosoft365MaterializeLightBlue PaletteMicrosoft365MaterializeLightBlue => _paletteMicrosoft365MaterializeLightBlue ??= new PaletteMicrosoft365MaterializeLightBlue();
+    public static PaletteMicrosoft365MaterializeLightBlueDark PaletteMicrosoft365MaterializeLightBlueDark => _paletteMicrosoft365MaterializeLightBlueDark ??= new PaletteMicrosoft365MaterializeLightBlueDark();
+    public static PaletteMicrosoft365SilverDarkModeAlternate PaletteMicrosoft365SilverDarkModeAlternate => _paletteMicrosoft365SilverDarkModeAlternate ??= new PaletteMicrosoft365SilverDarkModeAlternate();
+    public static PaletteMaterialMaterializeBlue PaletteMaterialMaterializeBlue => _paletteMaterialMaterializeBlue ??= new PaletteMaterialMaterializeBlue();
+    public static PaletteMaterialMaterializeBlueDark PaletteMaterialMaterializeBlueDark => _paletteMaterialMaterializeBlueDark ??= new PaletteMaterialMaterializeBlueDark();
+    public static PaletteMaterialMaterializeBlueRipple PaletteMaterialMaterializeBlueRipple => _paletteMaterialMaterializeBlueRipple ??= new PaletteMaterialMaterializeBlueRipple();
+    public static PaletteMaterialMaterializeBlueDarkRipple PaletteMaterialMaterializeBlueDarkRipple => _paletteMaterialMaterializeBlueDarkRipple ??= new PaletteMaterialMaterializeBlueDarkRipple();
+    public static PaletteMaterialMaterializeLightBlue PaletteMaterialMaterializeLightBlue => _paletteMaterialMaterializeLightBlue ??= new PaletteMaterialMaterializeLightBlue();
+    public static PaletteMaterialMaterializeLightBlueDark PaletteMaterialMaterializeLightBlueDark => _paletteMaterialMaterializeLightBlueDark ??= new PaletteMaterialMaterializeLightBlueDark();
+    public static PaletteMaterialMaterializeLightBlueRipple PaletteMaterialMaterializeLightBlueRipple => _paletteMaterialMaterializeLightBlueRipple ??= new PaletteMaterialMaterializeLightBlueRipple();
+    public static PaletteMaterialMaterializeLightBlueDarkRipple PaletteMaterialMaterializeLightBlueDarkRipple => _paletteMaterialMaterializeLightBlueDarkRipple ??= new PaletteMaterialMaterializeLightBlueDarkRipple();
+    public static PaletteMaterialSilverDarkModeAlternate PaletteMaterialSilverDarkModeAlternate => _paletteMaterialSilverDarkModeAlternate ??= new PaletteMaterialSilverDarkModeAlternate();
+    public static PaletteMaterialSilverDarkModeAlternateRipple PaletteMaterialSilverDarkModeAlternateRipple => _paletteMaterialSilverDarkModeAlternateRipple ??= new PaletteMaterialSilverDarkModeAlternateRipple();
+
+    private static PaletteMaterialLight? _paletteMaterialLight;
+    private static PaletteMaterialDark? _paletteMaterialDark;
+    private static PaletteMaterialLightRipple? _paletteMaterialLightRipple;
+    private static PaletteMaterialDarkRipple? _paletteMaterialDarkRipple;
+    private static PaletteMaterialLimeGreen? _paletteMaterialLimeGreen;
+    private static PaletteMaterialLimeGreenDark? _paletteMaterialLimeGreenDark;
+    private static PaletteMaterialLimeGreenRipple? _paletteMaterialLimeGreenRipple;
+    private static PaletteMaterialLimeGreenDarkRipple? _paletteMaterialLimeGreenDarkRipple;
+    private static PaletteMaterialDarkGray? _paletteMaterialDarkGray;
+    private static PaletteMaterialLightGray? _paletteMaterialLightGray;
+    private static PaletteMaterialDarkGrayRipple? _paletteMaterialDarkGrayRipple;
+    private static PaletteMaterialLightGrayRipple? _paletteMaterialLightGrayRipple;
+
+    private static PaletteRetroGreen? _paletteRetroGreen;
+    private static PaletteRetroBlue? _paletteRetroBlue;
+    private static PaletteMacOSXAqua? _paletteMacOSXAqua;
+    private static PaletteMacOSLight? _paletteMacOSLight;
+    private static PaletteMacOSDark? _paletteMacOSDark;
+    private static PaletteHighContrast? _paletteHighContrast;
+    private static PaletteDeuteranopia? _paletteDeuteranopia;
+    private static PaletteProtanopia? _paletteProtanopia;
+    private static PaletteOffice2007HighContrast? _paletteOffice2007HighContrast;
+    private static PaletteOffice2007Deuteranopia? _paletteOffice2007Deuteranopia;
+    private static PaletteOffice2007Protanopia? _paletteOffice2007Protanopia;
+    private static PaletteOffice2010HighContrast? _paletteOffice2010HighContrast;
+    private static PaletteOffice2010Deuteranopia? _paletteOffice2010Deuteranopia;
+    private static PaletteOffice2010Protanopia? _paletteOffice2010Protanopia;
+    private static PaletteOffice2013HighContrast? _paletteOffice2013HighContrast;
+    private static PaletteOffice2013Deuteranopia? _paletteOffice2013Deuteranopia;
+    private static PaletteOffice2013Protanopia? _paletteOffice2013Protanopia;
+    private static PaletteSparkleHighContrast? _paletteSparkleHighContrast;
+    private static PaletteSparkleDeuteranopia? _paletteSparkleDeuteranopia;
+    private static PaletteSparkleProtanopia? _paletteSparkleProtanopia;
+    private static PaletteMaterialHighContrast? _paletteMaterialHighContrast;
+    private static PaletteMaterialDeuteranopia? _paletteMaterialDeuteranopia;
+    private static PaletteMaterialProtanopia? _paletteMaterialProtanopia;
+    private static PaletteMaterialHighContrastRipple? _paletteMaterialHighContrastRipple;
+    private static PaletteMaterialDeuteranopiaRipple? _paletteMaterialDeuteranopiaRipple;
+    private static PaletteMaterialProtanopiaRipple? _paletteMaterialProtanopiaRipple;
+
+    private static PaletteOffice2007LimeGreen? _paletteOffice2007LimeGreen;
+    private static PaletteOffice2007LimeGreenDark? _paletteOffice2007LimeGreenDark;
+    private static PaletteOffice2010LimeGreen? _paletteOffice2010LimeGreen;
+    private static PaletteOffice2010LimeGreenDark? _paletteOffice2010LimeGreenDark;
+    private static PaletteMicrosoft365LimeGreen? _paletteMicrosoft365LimeGreen;
+    private static PaletteMicrosoft365LimeGreenDark? _paletteMicrosoft365LimeGreenDark;
+    private static PaletteOffice2007MaterializeBlue? _paletteOffice2007MaterializeBlue;
+    private static PaletteOffice2007MaterializeBlueDark? _paletteOffice2007MaterializeBlueDark;
+    private static PaletteOffice2007MaterializeLightBlue? _paletteOffice2007MaterializeLightBlue;
+    private static PaletteOffice2007MaterializeLightBlueDark? _paletteOffice2007MaterializeLightBlueDark;
+    private static PaletteOffice2007SilverDarkModeAlternate? _paletteOffice2007SilverDarkModeAlternate;
+    private static PaletteOffice2010MaterializeBlue? _paletteOffice2010MaterializeBlue;
+    private static PaletteOffice2010MaterializeBlueDark? _paletteOffice2010MaterializeBlueDark;
+    private static PaletteOffice2010MaterializeLightBlue? _paletteOffice2010MaterializeLightBlue;
+    private static PaletteOffice2010MaterializeLightBlueDark? _paletteOffice2010MaterializeLightBlueDark;
+    private static PaletteOffice2010SilverDarkModeAlternate? _paletteOffice2010SilverDarkModeAlternate;
+    private static PaletteOffice2013MaterializeBlue? _paletteOffice2013MaterializeBlue;
+    private static PaletteOffice2013MaterializeBlueDark? _paletteOffice2013MaterializeBlueDark;
+    private static PaletteOffice2013MaterializeLightBlue? _paletteOffice2013MaterializeLightBlue;
+    private static PaletteOffice2013MaterializeLightBlueDark? _paletteOffice2013MaterializeLightBlueDark;
+    private static PaletteOffice2013SilverDarkModeAlternate? _paletteOffice2013SilverDarkModeAlternate;
+    private static PaletteMicrosoft365MaterializeBlue? _paletteMicrosoft365MaterializeBlue;
+    private static PaletteMicrosoft365MaterializeBlueDark? _paletteMicrosoft365MaterializeBlueDark;
+    private static PaletteMicrosoft365MaterializeLightBlue? _paletteMicrosoft365MaterializeLightBlue;
+    private static PaletteMicrosoft365MaterializeLightBlueDark? _paletteMicrosoft365MaterializeLightBlueDark;
+    private static PaletteMicrosoft365SilverDarkModeAlternate? _paletteMicrosoft365SilverDarkModeAlternate;
+    private static PaletteMaterialMaterializeBlue? _paletteMaterialMaterializeBlue;
+    private static PaletteMaterialMaterializeBlueDark? _paletteMaterialMaterializeBlueDark;
+    private static PaletteMaterialMaterializeBlueRipple? _paletteMaterialMaterializeBlueRipple;
+    private static PaletteMaterialMaterializeBlueDarkRipple? _paletteMaterialMaterializeBlueDarkRipple;
+    private static PaletteMaterialMaterializeLightBlue? _paletteMaterialMaterializeLightBlue;
+    private static PaletteMaterialMaterializeLightBlueDark? _paletteMaterialMaterializeLightBlueDark;
+    private static PaletteMaterialMaterializeLightBlueRipple? _paletteMaterialMaterializeLightBlueRipple;
+    private static PaletteMaterialMaterializeLightBlueDarkRipple? _paletteMaterialMaterializeLightBlueDarkRipple;
+    private static PaletteMaterialSilverDarkModeAlternate? _paletteMaterialSilverDarkModeAlternate;
+    private static PaletteMaterialSilverDarkModeAlternateRipple? _paletteMaterialSilverDarkModeAlternateRipple;
 
     //public static PaletteBase CustomPaletteBase => _customPalette ??= new PaletteBase ();
 
@@ -2211,6 +2319,11 @@ public sealed class KryptonManager : Component
             case PaletteMode.Office2007BlackDarkMode:
             case PaletteMode.Office2007LimeGreen:
             case PaletteMode.Office2007LimeGreenDark:
+            case PaletteMode.Office2007MaterializeBlue:
+            case PaletteMode.Office2007MaterializeBlueDark:
+            case PaletteMode.Office2007MaterializeLightBlue:
+            case PaletteMode.Office2007MaterializeLightBlueDark:
+            case PaletteMode.Office2007SilverDarkModeAlternate:
             case PaletteMode.Office2007DarkGray:
             case PaletteMode.Office2007LightGray:
             case PaletteMode.VisualStudio2010Render2007:
@@ -2242,6 +2355,11 @@ public sealed class KryptonManager : Component
             case PaletteMode.SparkleProtanopia:
             case PaletteMode.Office2010LimeGreen:
             case PaletteMode.Office2010LimeGreenDark:
+            case PaletteMode.Office2010MaterializeBlue:
+            case PaletteMode.Office2010MaterializeBlueDark:
+            case PaletteMode.Office2010MaterializeLightBlue:
+            case PaletteMode.Office2010MaterializeLightBlueDark:
+            case PaletteMode.Office2010SilverDarkModeAlternate:
             case PaletteMode.Office2010DarkGray:
             case PaletteMode.Office2010LightGray:
             case PaletteMode.VisualStudio2010Render2010:
@@ -2257,6 +2375,11 @@ public sealed class KryptonManager : Component
             case PaletteMode.Office2013HighContrast:
             case PaletteMode.Office2013Deuteranopia:
             case PaletteMode.Office2013Protanopia:
+            case PaletteMode.Office2013MaterializeBlue:
+            case PaletteMode.Office2013MaterializeBlueDark:
+            case PaletteMode.Office2013MaterializeLightBlue:
+            case PaletteMode.Office2013MaterializeLightBlueDark:
+            case PaletteMode.Office2013SilverDarkModeAlternate:
                 Images.ToolbarImages.SetToolBarImages(ToolkitStaticVariables.Office2013ToolBarImages);
                 break;
             case PaletteMode.Microsoft365Black:
@@ -2271,6 +2394,11 @@ public sealed class KryptonManager : Component
             case PaletteMode.Microsoft365White:
             case PaletteMode.Microsoft365LimeGreen:
             case PaletteMode.Microsoft365LimeGreenDark:
+            case PaletteMode.Microsoft365MaterializeBlue:
+            case PaletteMode.Microsoft365MaterializeBlueDark:
+            case PaletteMode.Microsoft365MaterializeLightBlue:
+            case PaletteMode.Microsoft365MaterializeLightBlueDark:
+            case PaletteMode.Microsoft365SilverDarkModeAlternate:
             case PaletteMode.Microsoft365DarkGray:
             case PaletteMode.Microsoft365LightGray:
             case PaletteMode.VisualStudio2010Render365:
@@ -2309,6 +2437,16 @@ public sealed class KryptonManager : Component
             case PaletteMode.MaterialLimeGreenDark:
             case PaletteMode.MaterialLimeGreenRipple:
             case PaletteMode.MaterialLimeGreenDarkRipple:
+            case PaletteMode.MaterialMaterializeBlue:
+            case PaletteMode.MaterialMaterializeBlueDark:
+            case PaletteMode.MaterialMaterializeBlueRipple:
+            case PaletteMode.MaterialMaterializeBlueDarkRipple:
+            case PaletteMode.MaterialMaterializeLightBlue:
+            case PaletteMode.MaterialMaterializeLightBlueDark:
+            case PaletteMode.MaterialMaterializeLightBlueRipple:
+            case PaletteMode.MaterialMaterializeLightBlueDarkRipple:
+            case PaletteMode.MaterialSilverDarkModeAlternate:
+            case PaletteMode.MaterialSilverDarkModeAlternateRipple:
             case PaletteMode.MaterialDarkGray:
             case PaletteMode.MaterialLightGray:
             case PaletteMode.MaterialDarkGrayRipple: